JVM IR: uninherit IrExpression from IrAttributeContainer
`IrExpression.attributeOwnerId` was only used in JVM IR for
IrFunctionReference and IrPropertyReference in several places, the main
of which was to store the names of local classes that are generated for
function/property reference classes. It is probably easier, and it saves
a field for all expressions, to keep these names explicitly in a new map
in JvmBackendContext and transfer them to the main map in the
corresponding lowerings.